+++ resources/views/admin/category/index.php
@@ -12,11 +12,17 @@ declare(strict_types=1);
 use Hirtz\Cms\Modules\Admin\Data\CategoryActiveDataProvider;
 use Hirtz\Cms\Modules\Admin\Widgets\Grids\CategoryGridView;
 use Hirtz\Cms\Modules\Admin\Widgets\Navs\CategoryHeader;
+use Hirtz\Cms\Modules\Admin\Widgets\Navs\CategorySubmenu;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Grids\GridContainer;
 
 echo CategoryHeader::make()
     ->provider($provider);
+
+if ($provider->category) {
+    echo CategorySubmenu::make()
+        ->model($provider->category);
+}
 
 echo GridContainer::make()
     ->grid(CategoryGridView::make()